requirements.txt
streamlit==1.28.0
python-dotenv==1.0.0
requests==2.31.0